requirements.txt
asgiref==3.11.0
Django==5.2.9
django-filter==25.2
django-phonenumber-field==8.4.0
django-resized==1.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
iniconfig==2.3.0
packaging==25.0
phonenumbers==9.0.21
pillow==12.1.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
